# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Sorteio Ativo",
    "version": "1.0.0",
    "description": "Documentação da API do Sorteio Ativo"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/admin/products/{id}": {
      "put": {
        "summary": "Atualiza um produto existente (Admin)",
        "description": "(Admin) Altera os detalhes de um produto específico.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "O ID numérico do produto a ser atualizado.",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Nova Rifa Incrível"
                  },
                  "description": {
                    "type": "string",
                    "example": "Uma descrição atualizada para a rifa."
                  },
                  "price_per_number": {
                    "type": "number",
                    "format": "float",
                    "example": 15.75
                  },
                  "image_url": {
                    "type": "string",
                    "example": "https://example.com/nova_imagem.jpg"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "upcoming",
                      "drawn",
                      "cancelled"
                    ],
                    "example": "active"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Produto atualizado com sucesso"
                    },
                    "product": {
                      "type": "object",
                      "description": "Retorna o produto atualizado"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados de entrada inválidos."
          },
          "401": {
            "description": "Não autenticado."
          },
          "403": {
            "description": "Acesso negado, não é um administrador."
          },
          "404": {
            "description": "Produto não encontrado."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
    "/api/admin/products": {
      "post": {
        "summary": "Cria um novo produto (sorteio)",
        "description": "(Admin) Adiciona um novo produto e inicializa seus números de 0 a 100 como disponíveis.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "price_per_number"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Rifa de um PlayStation 5"
                  },
                  "description": {
                    "type": "string",
                    "example": "Concorra a um PS5 novinho em folha!"
                  },
                  "price_per_number": {
                    "type": "number",
                    "format": "float",
                    "example": 25.5
                  },
                  "image_url": {
                    "type": "string",
                    "example": "https://example.com/ps5.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto e números criados com sucesso."
          },
          "401": {
            "description": "Não autenticado."
          },
          "403": {
            "description": "Acesso negado, não é um administrador."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      },
      "get": {
        "summary": "Lista todos os produtos (Admin)",
        "description": "(Admin) Retorna uma lista de todos os produtos cadastrados, independentemente do status.\n",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Uma lista de todos os produtos.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "price_per_number": {
                            "type": "number",
                            "format": "float"
                          },
                          "image_url": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "active",
                              "upcoming",
                              "drawn",
                              "cancelled"
                            ]
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "winning_number": {
                            "type": "integer",
                            "nullable": true
                          },
                          "winner_user_id": {
                            "type": "integer",
                            "nullable": true
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado."
          },
          "403": {
            "description": "Acesso negado, não é um administrador."
          },
          "500": {
            "description": "Erro interno do servidor."
          }
        }
      }
    },
...
  },
...
}